Extract PMA thread commands into dedicated module with registration hook
Move all 13 thread command functions, 6 dataclasses, and ~15 helper functions
from pma_cli.py into pma_thread_commands.py with a register_thread_commands()
hook. Migrate 41 thread-specific tests into test_pma_thread_commands.py and
update monkeypatch/patch targets accordingly.
