+++ pyproject.toml
@@ -127,8 +127,9 @@ ignore = ["E501"]
 "src/codex_autorunner/pty_session.py" = ["F403"]
 "src/codex_autorunner/static_assets.py" = ["F403"]
 "src/codex_autorunner/cli.py" = ["B008"]
-"src/codex_autorunner/surfaces/cli/cli.py" = ["B008"]
 "src/codex_autorunner/surfaces/cli/pma_cli.py" = ["B008"]
+"src/codex_autorunner/surfaces/cli/pma_thread_commands.py" = ["B008"]
+"src/codex_autorunner/surfaces/cli/cli.py" = ["B008"]
 "src/codex_autorunner/surfaces/cli/commands/*.py" = ["B008"]
 "src/codex_autorunner/surfaces/web/routes/terminal_images.py" = ["B008"]
 "src/codex_autorunner/surfaces/web/routes/voice.py" = ["B008"]
